src/requirements.txt
matplotlib~=3.3.2
pandas~=1.1.3
psycopg2~=2.8.6
seaborn~=0.11.0
numpy~=1.19.2
sklearn~=0.0
scikit-learn~=0.23.2
scipy~=1.5.2
python-dotenv~=0.15.0